+++ endhost/ssp/SCIONProtocol.cpp
@@ -301,14 +301,7 @@ int SSPProtocol::recv(uint8_t *buf, size_t len, SCIONAddr *srcAddr, double timeo
             return -EWOULDBLOCK;
         }
         if (timeout > 0.0) {
-            struct timespec ts;
-            struct timeval tv;
-            int secs = (int)timeout;
-            uint64_t ns = (timeout - secs) * 1000000000;
-            gettimeofday(&tv, NULL);
-            ts.tv_sec = tv.tv_sec + (int)timeout;
-            ts.tv_nsec = tv.tv_usec * 1000 + ns;
-            if (pthread_cond_timedwait(&mReadCond, &mReadMutex, &ts) == ETIMEDOUT) {
+            if (timedWait(&mReadCond, &mReadMutex, timeout) == ETIMEDOUT) {
                 pthread_mutex_unlock(&mReadMutex);
                 DEBUG("%p: timeout in recv\n", this);
                 return -ETIMEDOUT;
